[FIX] change size of counter to u64 else it can be to big

+++ client/src/hash_cash.rs
@@ -6,7 +6,7 @@ use shared::md5_hash_cash_output::MD5HashCashOutput;
 pub struct HashCash {
     pub input: MD5HashCashInput,
     pub output: MD5HashCashOutput,
-    counter: u128,
+    counter: u64,
     is_valid: bool,
 }
